# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "GeneSilico EHR API",
        "summary": "Ontada (iKnowMed) FHIR + Stedi eligibility for Texas Oncology.",
...
        "contact": {
            "name": "GeneSilico",
            "email": "developers@genesilico.ai"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8090",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/eligibility/form": {
            "get": {
                "tags": [
                    "eligibility"
                ],
                "summary": "Get Form",
                "description": "The payer-specific field schema the UI renders.",
                "operationId": "get_form_api_eligibility_form_get",
                "parameters": [
                    {
                        "name": "insurer",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "Payer name as the user typed it",
                            "default": "",
                            "title": "Insurer"
                        },
                        "description": "Payer name as the user typed it"
                    },
                    {
                        "name": "payer_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "Stedi trading partner service id",
                            "default": "",
                            "title": "Payer Id"
                        },
                        "description": "Stedi trading partner service id"
                    },
                    {
                        "name": "plan_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "",
                            "title": "Plan Type"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PolicyForm"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "428": {
                        "description": "A precondition is unmet \u2014 typically Ontada is not connected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Deliberately not implemented; the reason is in the message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "An upstream (Stedi, Ontada) failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "A capability is not configured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/eligibility/run": {
            "post": {
                "tags": [
                    "eligibility"
                ],
                "summary": "Run",
                "description": "Validate the form, then either call the payer or replay a stored 271.",
                "operationId": "run_api_eligibility_run_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RunRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EligibilityRun"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "428": {
                        "description": "A precondition is unmet \u2014 typically Ontada is not connected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Deliberately not implemented; the reason is in the message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "An upstream (Stedi, Ontada) failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "A capability is not configured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EligibilityRun": {
                "properties": {
                    "source": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Source",
                        "description": "{\"kind\":\"live\"|\"replay\", \u2026}"
                    },
                    "benefits": {
                        "$ref": "#/components/schemas/ParsedBenefits"
                    },
                    "estimate": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Estimate"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "source",
                    "benefits"
                ],
                "title": "EligibilityRun"
            },
            "ErrorBody": {
                "properties": {
                    "code": {
                        "type": "string",
                        "title": "Code",
                        "description": "Machine-readable: unconfigured, upstream_error, ontada_not_connected, form_invalid, not_implemented \u2026"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "Human-readable. Upstream messages are kept verbatim."
                    },
                    "detail": {
                        "anyOf": [
                            {},
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": [
                    "code",
                    "message"
                ],
                "title": "ErrorBody"
            },
            "ErrorEnvelope": {
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorBody"
                    }
                },
                "type": "object",
                "required": [
                    "error"
                ],
                "title": "ErrorEnvelope",
                "description": "Every non-2xx response from this API has exactly this shape."
            },
            "Estimate": {
                "properties": {
                    "lines": {
                        "items": {
                            "$ref": "#/components/schemas/PreAuthLineOut"
                        },
                        "type": "array",
                        "title": "Lines"
                    },
                    "network": {
                        "type": "string",
                        "title": "Network"
                    },
                    "oop_remaining": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Oop Remaining",
                        "description": "Caps the running total. A real figure from the payer."
                    },
                    "totals": {
                        "$ref": "#/components/schemas/EstimateTotals"
                    },
                    "disclaimer": {
                        "type": "string",
                        "title": "Disclaimer"
                    }
                },
                "type": "object",
                "required": [
                    "lines",
                    "network",
                    "totals",
                    "disclaimer"
                ],
                "title": "Estimate"
            },
...
            "FormField": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "label": {
                        "type": "string",
                        "title": "Label"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "text",
                            "select",
                            "date",
                            "checkbox",
                            "tel",
                            "number"
                        ],
                        "title": "Type"
                    },
                    "required": {
                        "type": "boolean",
                        "title": "Required",
                        "default": false
                    },
                    "value": {
                        "anyOf": [
                            {},
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Value"
                    },
                    "options": {
                        "anyOf": [
                            {
                                "items": {
                                    "additionalProperties": true,
                                    "type": "object"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Options"
                    },
                    "help": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Help"
                    },
                    "pattern": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Pattern"
                    },
                    "maxLength": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Maxlength"
                    },
                    "visible_when": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Visible When",
                        "description": "Show only when these values hold"
                    },
                    "sensitive": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Sensitive"
                    },
                    "x12": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "X12",
                        "description": "The X12 element this field maps onto \u2014 the contract between the form and the 270 it produces"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "label",
                    "type"
                ],
                "title": "FormField"
            },
            "FormSection": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "fields": {
                        "items": {
                            "$ref": "#/components/schemas/FormField"
                        },
                        "type": "array",
                        "title": "Fields"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "title",
                    "fields"
                ],
                "title": "FormSection"
            },
...
            "ParsedBenefits": {
                "properties": {
                    "active": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Active"
                    },
                    "payer": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Payer"
                    },
                    "plan": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Plan"
                    },
                    "member": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Member"
                    },
                    "accumulators": {
                        "items": {
                            "$ref": "#/components/schemas/Accumulator"
                        },
                        "type": "array",
                        "title": "Accumulators"
                    },
                    "drug_tiers": {
                        "items": {
                            "$ref": "#/components/schemas/DrugTier"
                        },
                        "type": "array",
                        "title": "Drug Tiers"
                    },
                    "coinsurance": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Coinsurance"
                    },
                    "limitations": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Limitations"
                    },
                    "referrals": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Referrals"
                    },
                    "errors": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Errors",
                        "description": "The payer's own AAA rejections, verbatim"
                    },
                    "trace_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Trace Id"
                    },
                    "mode": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Mode"
                    },
                    "entry_count": {
                        "type": "integer",
                        "title": "Entry Count"
                    }
                },
                "type": "object",
                "required": [
                    "active",
                    "payer",
                    "plan",
                    "member",
                    "accumulators",
                    "drug_tiers",
                    "coinsurance",
                    "limitations",
                    "referrals",
                    "errors",
                    "trace_id",
                    "mode",
                    "entry_count"
                ],
                "title": "ParsedBenefits",
                "description": "A real 271 reduced to what a pre-auth screen needs.\n\nA commercial plan can carry 700+ benefit lines with no coinsurance and no\ndeductible at all \u2014 oncology is priced by drug tier. Hence `drug_tiers`."
            },
...
            "PolicyForm": {
                "properties": {
                    "form_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Form Id"
                    },
                    "form_version": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Form Version"
                    },
                    "family": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Family",
                        "description": "{id, label, why} \u2014 which plan-family rules apply"
                    },
                    "payer": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Payer"
                    },
                    "sections": {
                        "items": {
                            "$ref": "#/components/schemas/FormSection"
                        },
                        "type": "array",
                        "title": "Sections"
                    },
                    "prefill": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Prefill",
                        "description": "Values already known, and where each came from, so a prefilled field is never mistaken for one the clinician typed"
                    },
                    "consent_text": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Consent Text",
                        "description": "Attestation the user must accept before a real payer call is made on this patient"
                    },
                    "sources": {
                        "anyOf": [
                            {
                                "items": {
                                    "additionalProperties": true,
                                    "type": "object"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Sources",
                        "description": "Citations for the payer-specific rules this form enforces (CMS, X12, BlueCard \u2026)"
                    }
                },
                "type": "object",
                "required": [
                    "family",
                    "sections"
                ],
                "title": "PolicyForm",
                "description": "Field schema for one payer.\n\nA Medicare MBI, a BlueCard alpha prefix, a TRICARE sponsor SSN and a\ncommercial member ID are not interchangeable. `family` says which set of\nrules applies and `why` says how it was decided."
            },
...
            "RunRequest": {
                "properties": {
                    "insurer": {
                        "type": "string",
                        "title": "Insurer",
                        "default": ""
                    },
                    "payer_id": {
                        "type": "string",
                        "title": "Payer Id",
                        "default": ""
                    },
                    "plan_type": {
                        "type": "string",
                        "title": "Plan Type",
                        "default": ""
                    },
                    "values": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Values"
                    },
                    "service_type_codes": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Service Type Codes"
                    },
                    "lines": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Lines"
                    },
                    "network": {
                        "type": "string",
                        "title": "Network",
                        "default": "in"
                    },
                    "replay": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Replay"
                    }
                },
                "type": "object",
                "title": "RunRequest"
            },
...
        }
    },
...
}